Fade between two Literal Values
By introducing fade over times we are forced to abstract the concept of
a Generator to a Trait. This requires it be stored as a Box Pointer in
the fixture, and allocate it to the heap. But in the case that all these
generators are the same, they can all point to the same generator -
which is nice.

+++ lumen/src/value.rs
@@ -1,10 +1,10 @@
 #[derive(Clone, Copy, Debug)]
 pub struct Literal {
-    pub value: i32,
+    pub value: f32,
 }
 
 impl Literal {
-    pub fn new(value: i32) -> Self {
+    pub fn new(value: f32) -> Self {
         Self { value }
     }
 }
